input をモジュールに分割
